projects/lupsEdgeServer/requirements.txt
# Install with: `pip install -r requirements.txt`

#core
bottle
apscheduler

# Infrastructure
Django >= 1.9.5, < 1.10
#Pillow >= 3.2.0
#for MySQL: MySQL-python >= 1.2.5
#for PostgreSQL: psycopg2 >= 2.6.1
#python-memcached >= 1.57
#raven >= 5.17.0
#uWSGI >= 2.0.13.1

# API REST
djangorestframework >=3.3.3
Markdown >= 2.1.0
django-filter >= 0.9.2
#django-crispy-forms
django-guardian >= 1.1.1
django-url-filter

# CMS libraries
#django-fluent-dashboard >= 1.0a1

# Support libs
#django-analytical >= 2.0.0
#django-admin-tools >= 0.7.2
#django-axes >= 1.6.1
#django-compressor >= 2.0
#django-crispy-forms >= 1.6.0
#-e git+https://github.com/smacker/django-filebrowser-no-grappelli.git@121c1bcf2c859e8dec8cf43040377ba23e16745b#egg=django-filebrowser-no-grappelli
-e git+https://github.com/rizumu/django-ping.git@2ff3985c762654282ba52813043a605b73896654#egg=django_ping-0.3b1
pytz >= 2016.4
#sorl-thumbnail >= 12.3
requests
paho-mqtt


# Debugging
django-debugtools >= 1.6
#ipython >= 4.2.0
#ptpython >= 0.34



#chardet==2.3.0
#cloud-init==0.7.7
#command-not-found==0.3
#configobj==5.0.6
#Jinja2==2.7.3
#jsonpatch==1.3
#jsonpointer==1.0
#language-selector==0.1
#MarkupSafe==0.23
#oauthlib==0.6.1
#pbr==1.9.1
#prettytable==0.7.2
#pycrypto==2.6.1
#pycurl==7.19.5
#Pygments==2.1.1
#pygobject==3.14.0
#pyserial==2.6
#python-apt===0.9.3.11build1
#PyYAML==3.11
#requests==2.4.3
#six==1.9.0
#ssh-import-id==4.1
#stevedore==1.12.0
#ufw===0.34-rc-0ubuntu5
#unattended-upgrades==0.1
#urllib3==1.9.1
#virtualenv==15.0.1
#virtualenv-clone==0.2.6
#virtualenvwrapper==4.7.0
